requirements.txt
Flask==3.0.3
mysql-connector-python==8.4.0
python-dotenv==1.0.1
gunicorn==21.2.0
